Refactor Storage and UI in package
