+++ src/main/java/modules/Storage.java
@@ -1,4 +1,4 @@
-package seedu.rainyDay;
+package modules;
 
 import seedu.rainyDay.data.FinancialReport;
 import seedu.rainyDay.data.FinancialStatement;
