+++ src/test/java/seedu/rainyDay/StorageTest.java
@@ -1,5 +1,6 @@
 package seedu.rainyDay;
 
+import modules.*;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import seedu.rainyDay.command.Command;
